refactor(Rust): Create `parser-r` crate
